Refactor Acme plugin into AcmeChallenge objects, dns01
